Use exception chaining & str(exc) (#255)
* Use exception chaining & str(exc)

Use exception chaining (via raise_from which is a stub on python2) so
that the entire traceback is captured by the exception.

Explicitly coerce exceptions to str() in error messages so we only get
the exception message and not the entire representation (stacktrace
etc).

This should mostly bring py2/py3 error logging behavior back into
sync, with the main difference that SourceLine only adds tracebacks to the exception message on PY2.  It is unnecessary to do this on PY3 because the entire exception chain and tracebacks are available.