Convert URIs to avro namespaces for types.
The key difference is that field names, symbols and type names are
unified in the Salad/RDF data model, but are separate in the Avro data
model.  Types and fields names and enum symbols need to have their
names mapped differently.  This is now handled.

The practical effect is that this fixes a longstanding issue in
referencing types from imported schemas, as is done in cwltool
extension schema to reference cwl:Expression.

refs #326