Fix support for 'subscope' in python codegen
The way subscopes are supposed to work is to amend the base URI used
for loading that subtree of the document.  The implementation just
changed the name of the variable (?)

This is fixed for Python.

The TypesSript, java and dotnet codegens are only fixed so that they
arn't broken and MyPy won't complain, but I did not actually add the
equivalent codegen.