Always re-run unittests
Currently running `tox -e py310-unit` will only run the unit tests the first time the command is run. 

This is because `make coverage-report` (the command that tox runs for the unit tests) checks to see .coverage exists. However, .coverage is created the first time and not deleted so the tests will not be rerun. The simplest solution with the least amount of potential downstream issues would be to add a `--always-make` flag in the tox.ini file.