Adding `map` and named `union` types to Schema SALAD (#672)
* Adding CWLObjectType named union
* Added `CWLInputFile` type
* Using UnionLoader for SecondaryFilesDSL
* Added NamedMapSchema and NamedUnionSchema
* Document Schema Salad 1.3
* Add support for `noLinkCheck` in codegen

This commit introduces support for the `noLinkCheck` `jsonldPredicate`
in the Schema SALAD codegen toolchain, and in particular in the
`URILoader` stack. When `noLinkCheck` is set to `true`, no link checking
is performed for all the underlying objects hierarchy. The need to
propagate its value to the hierarchy makes it necessary to modify also
the `LoadingOptions` data structure.

All parsers but the Python one never perform link checking up to now, so
this update makes no modification to their actual behaviour. However, it
enables the `URILoader` classes to receive the `noLinkCheck` parameter
for future implementations.

* Updated typescript generator
* Updated cpp generator
* Fix misinterpretation of absent fields

This commit removes an erroneous null check, which caused parsers to
evaluate as erroneous some correct documents.

* Updated CWL schema tests
* Make `@list` container preserve nesting
* Propagate `noLinkCheck` through maps and records