feat(rust): add rust codegen module
This commit includes the Python module responsible for generating code
based on a given schema.

The `schema_salad/rust/salad-core` folder serves as a placeholder for
Rust code logic that will enable document (de)serialization, which is
still a work in progress due to ongoing code refactoring.
The code for this refactoring can be found in the
"giuseppe998e/schema_salad" repository branch.